[hipDNN] 修复隐式 size_t→double 转换编译错误
CI 的 -Wimplicit-int-float-conversion + -Werror 拒绝了
sumAbsDiff / summary.mismatchCount 的隐式转换。
改为显式 static_cast<double>。

+++ projects/hipdnn/test_sdk/include/hipdnn_test_sdk/utilities/TensorDiff.hpp
@@ -122,7 +122,8 @@ TensorDiffSummary computeTensorDiff(hipdnn_data_sdk::utilities::ITensor& referen
 
     if(summary.mismatchCount > 0)
     {
-        summary.meanAbsDiff = static_cast<float>(sumAbsDiff / summary.mismatchCount);
+        summary.meanAbsDiff
+            = static_cast<float>(sumAbsDiff / static_cast<double>(summary.mismatchCount));
 
         std::sort(summary.worstMismatches.begin(),
                   summary.worstMismatches.end(),
